+++ src/client/hud/GameRenderer.ts
@@ -5,6 +5,7 @@ import { AmbienceController } from "../controllers/AmbienceController";
 import { AttackingTroopsController } from "../controllers/AttackingTroopsController";
 import { BuildPreviewController } from "../controllers/BuildPreviewController";
 import { HoverHighlightController } from "../controllers/HoverHighlightController";
+import { HudVisibilityController } from "../controllers/HudVisibilityController";
 import { ImpactFeedbackController } from "../controllers/ImpactFeedbackController";
 import { LiveStatsController } from "../controllers/LiveStatsController";
 import { MapLayerController } from "../controllers/MapLayerController";
@@ -344,6 +345,7 @@ export function createRenderer(
     new AttackingTroopsController(game, eventBus, userSettings, view),
     new SoundEffectController(game, eventBus),
     new AmbienceController(game, eventBus, transformHandler),
+    new HudVisibilityController(eventBus, userSettings),
     new ImpactFeedbackController(game, transformHandler, view),
     ...(mapLayerController ? [mapLayerController] : []),
     eventsDisplay,